Extract viewportFeatureClass generator
